Rename files to be more consistent
don't need `_jax` at the end of everything.
