+++ jax_lensing_profiles/MassModel/Profiles/__init__.py
@@ -1,8 +1,8 @@
-from .gaussian_ellipse_kappa_jax import GaussianEllipseKappa
-from .gaussian_kappa_jax import GaussianKappa
-from .MGE_jax import MGE
-from .multi_gaussian_ellipse_kappa_jax import MultiGaussianEllipseKappa
+from .gaussian_ellipse_kappa import GaussianEllipseKappa
+from .gaussian_kappa import GaussianKappa
+from .MGE import MGE
+from .multi_gaussian_ellipse_kappa import MultiGaussianEllipseKappa
 from .NFW import NFW
 from .NFW import TNFW
-from .NFW_ellipse_kappa_jax import NFWEllipseKappa
-from .Sersic_ellipse_kappa_jax import SersicEllipseKappa
+from .NFW_ellipse_kappa import NFWEllipseKappa
+from .Sersic_ellipse_kappa import SersicEllipseKappa
