+++ jax_lensing_profiles/MassModel/Profiles/gaussian_ellipse_kappa.py
