+++ jax_lensing_profiles/MassModel/Profiles/gaussian_kappa.py
