+++ jax_lensing_profiles/MassModel/Profiles/multi_gaussian_ellipse_kappa.py
